Add login and signup flow
We can customize the SignupViewController to take in the additional
fields we need, right now it is just using the default PFUser object
and treating "username" as "email". When we get design assets for how
this should look, we can customize the fields. We don't need to use
the parse default SignupViewController, it should be really easy to
implement our own and use the delegate.